changed the type for variable sum

+++ 0x10-variadic_functions/0-sum_them_all.c
@@ -9,11 +9,12 @@
 int sum_them_all(const unsigned int n, ...)
 {
 	va_list num_list;
-	double sum = 0.0;
+	int sum = 0;
 	unsigned int i;
 
 	if (n == 0)
 		return (0);
+
 	va_start(num_list, n);
 
 	for (i = 0; i < n; i++)
